requirements.txt
frappe>=15.0.0
erpnext>=15.0.0
Pillow>=9.0.0
pytesseract>=0.3.10
opencv-python>=4.5.0
python-magic>=0.4.27
cryptography>=3.4.8
elasticsearch>=7.0.0
redis>=4.0.0
celery>=5.0.0
requests>=2.25.0
python-dateutil>=2.8.0
pandas>=1.3.0
numpy>=1.21.0